build: refactor grass.py in preparation for FHS
(This is extracted from #5630 to simplify review.)

The transitional period from the traditional GRASS installation
structure to a Filesystem Hierarchy Standard (FHS) complying structure
(enabled by CMake build) requires an alternative way to find resources.
Currently all resources are located in relation to GISBASE, that is no
longer the case with a FHS installation.

In addition, and instead of GISBASE, the following environment
variables will be needed (not included in this update):

- GRASS_SHAREDIR
- GRASS_LOCALEDIR
- GRASS_PYDIR
- GRASS_GUIWXDIR
- GRASS_GUISCRIPTDIR
- GRASS_GUIRESDIR
- GRASS_FONTSDIR
- GRASS_ETCDIR

Setting up all these variables would bloat the grass.py file, therefore
this update suggests to move this to the grass.app Python module.

Summary of changes:

- grass.py: to simplify Conda adoption, the variable GRASS_PREFIX is
  added, which is essentially equal to installation prefix, but easily
  changeable if needed.
- grass.py need to find the Python module to be able to set necessary
  variables, a new variable GRASS_PYDIR is configured into the file to
  do just this.
- python/grass/app/resource_paths.py is a configurable file (like
  grass.py)
- update copy_python_files_in_subdir.cmake to be able to exclude files